+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # pixeltileengine changelog
 
+## 1.2.0 - 2026-01-14
+
+### Water shimmer and refraction effetcs
+
+### Fixed
+
+- Many of the particle effects were fixed or improved
+- Fixes to scaling and rotation
+
 ## 1.1.0 - 2025-04-19
 
 ### Added